Week 3
This week, I changed the grid to a 4x4 layout to make the game more
interactive and added action listeners to handle clicks on each card
which allows users to actually interact with the game.
To improve the experience, I implemented logic to prevent actions if
cards are already flipped/matched, and now matched cards display as
“Matched” and are grayed out. When the player finds a match, they earn 1
point, and the move counter is adjusted accordingly.
I also added a 1000 millisecond (1 second) timer to unflip unmatched
cards so that the user can see the card before they flip back. The same
method used to do that was used for the show all cards button, which
displays all cards.
I also implemented a restart button that reshuffles the cards and resets
the game state. Overall, I’m happy with the progress made so far. Next
week, I plan to add colors to the GUI to make the game more
aesthetically pleasing as well as make different card types that give
different points.