Week 4: This week I fixed a major bug within my code. When I click the same card twice, it matches the card with itself and is matched with solved. To address it I implemented logic to prevent users from clicking a card that is already flipped and matched by giving it the flipped property. I also started creating the different classes to incorporate inheritance and polymorphism. These cards include SpellCard, TroopCard, BuildingCard, which are all different colors and give a different number of points. I plan to add a lot more classes later on to make it the full 4x4 layout. Here is my GUI so far:
I’ll probably make it so that the text on each card changes to “Solved”
when they are paired.
To keep things simpler, I made it a 2x3 layout. This makes it so I won’t
have to implement 8 classes right away and can just stick to 3 for
testing purposes. I plan to turn it back to 4x4 in the future.
I also added some feedback when the player wins. It works by checking if
any of the cards on the grid are not matched, and if none are then a
popup shows up:

I’m pleased with my project so far. In the future, I plan to implement
all 8 classes, make matched pairs text changed to “solved”, and do
various GUI improvements.