Week 7: This week, I did a few minor debugging/adjustments. First of all, I made it so when you click the “Show Solution” button, you cannot click anything else. This is very good for making it so the player isn’t able to win the game by simply clicking the “Show Solution” button and then solving everything while seeing all the matches.
I also made it so the “Moves” variable updates every time 2 cards are
clicked, not every time one card is clicked. This makes it so the moves
variable accurately reflects how many matches the player has attempted,
not how many cards the player has clicked.
I also debugged a problem where if you clicked 2 cards then clicked the
“Show Solution” button while the timer was still active, the 2 cards you
clicked would revert back to their original state. I fixed this by
adding a boolean “isTimerActive” variable, which tracks if a timer is
running. When the timer is active, the “Show Solution” button checks the
status of the “isTimerActive” variable and halts timers if there are
timers running. Cards are no longer reset when the “Show Solution”
button is clicked.